# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amrutam Telemedicine API",
    "version": "1.0.0",
    "description": "Backend for consultations, doctor availability and prescriptions.\n\n## Idempotency\nEvery non-idempotent write requires an `Idempotency-Key` header. Retrying with the\nsame key replays the original response rather than repeating the side effect. Keys\nare scoped per user and expire after 24 hours.\n\n## Errors\nEvery error is `{ \"error\": { \"code\", \"message\", \"requestId\" } }`. Quote `requestId`\nin a support request — it correlates the response with logs and traces.\n\n## Rate limits\nResponses carry `RateLimit-Limit` and `RateLimit-Remaining`. Exceeding a limit returns\n429 with `Retry-After`.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /auth/login. Expires in 15 minutes; rotate with /auth/refresh."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "SLOT_UNAVAILABLE"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/doctors": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Search and filter doctors",
        "description": "Public. Results are cached for 60 seconds; only verified, active doctors are returned.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "required": false,
            "name": "specialization",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 32
            },
            "required": false,
            "name": "language",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 80
            },
            "required": false,
            "name": "city",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "maximum": 70
            },
            "required": false,
            "name": "minExperience",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            },
            "required": false,
            "name": "maxFeePaise",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "rating",
                "fee_asc",
                "fee_desc",
                "experience"
              ],
              "default": "rating"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching doctors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {}
                        },
                        "total": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "items",
                        "total"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/{id}": {
      "get": {
        "tags": [
          "Doctors"
        ],
        "summary": "Doctor profile",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/{id}/slots": {
      "get": {
        "tags": [
          "Doctors",
          "Availability"
        ],
        "summary": "Bookable slots for a date range",
        "description": "Cached for 30 seconds. A slot listed here can still be taken before you book it — expect 409.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true,
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Available slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {}
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/profile": {
      "post": {
        "tags": [
          "Doctors"
        ],
        "summary": "Create the calling doctor's profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registrationNo": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 64
                  },
                  "specializations": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 64
                    },
                    "minItems": 1,
                    "maxItems": 10
                  },
                  "languages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 2,
                      "maxLength": 32
                    },
                    "minItems": 1,
                    "maxItems": 10
                  },
                  "yearsExperience": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 70
                  },
                  "consultationFeePaise": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000
                  },
                  "city": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "registrationNo",
                  "specializations",
                  "languages",
                  "yearsExperience",
                  "consultationFeePaise"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created, pending verification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not a doctor account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Profile already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/me/availability/rules": {
      "post": {
        "tags": [
          "Availability"
        ],
        "summary": "Add a recurring weekly availability rule",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "weekday": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 6
                  },
                  "startTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "endTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "slotMinutes": {
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 240,
                    "default": 30
                  },
                  "timezone": {
                    "type": "string",
                    "default": "Asia/Kolkata"
                  },
                  "validFrom": {
                    "type": "string",
                    "format": "date"
                  },
                  "validTo": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "weekday",
                  "startTime",
                  "endTime"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rule created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "id"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/me/availability/generate": {
      "post": {
        "tags": [
          "Availability"
        ],
        "summary": "Materialise slots from rules for a date range",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Naturally idempotent: re-running over the same window creates no duplicates.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": {
                    "type": "string",
                    "format": "date"
                  },
                  "to": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "from",
                  "to"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Slots created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "slotsCreated": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "slotsCreated"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/doctors/{id}/verify": {
      "post": {
        "tags": [
          "Doctors",
          "Admin"
        ],
        "summary": "Verify a doctor (admin, MFA required)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Verified",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {}
                  }
                }
              }
            }
          },
          "401": {
            "description": "MFA required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}